Refactor: import from siblings not ../functions.ts
